refactor: 예외 발생 시 errorDto(ExceptionResponse) 응답
